feature: configurable routing table
the configuration is in :kv, :routing_table
the default value is an empty []
in config/config.exs we set it to be a single entry for all buckets,
with the current node
for our routing test, we dynamically change it to two entries,
named `a-to-m` and `n-to-z`. This test is tagged @distributed, and for
it to succeed two :kv nodes must be running with those names.